chore: project structure
