+++ src/components/table/table.utils.ts
@@ -1,13 +1,12 @@
-import type { ActivityRow } from "@/domains/activity/activity.types";
 import type { Header, SortDirection } from "@tanstack/react-table";
 
-export function getHeadCellSortDirection(header: Header<ActivityRow, unknown>) {
+export function getHeadCellSortDirection<TData>(header: Header<TData, unknown>) {
   return header.column.getIsSorted() === false
     ? undefined
     : (header.column.getIsSorted() as SortDirection);
 }
 
-export function getHeadCellSortTitle(header: Header<ActivityRow, unknown>) {
+export function getHeadCellSortTitle<TData>(header: Header<TData, unknown>) {
   return header.column.getCanSort()
     ? header.column.getNextSortingOrder() === "asc"
       ? "Sort by ascending order"
